requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bson==0.5.10
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.100.1
h11==0.14.0
httptools==0.6.4
idna==3.10
jwt==1.3.1
mangum==0.19.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.23.2
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
